refactor : 빈을 자동 스캔하는 기능으로 변환
