+++ src/main/java/roomescape/RoomescapeApplication.java
@@ -1,12 +1,11 @@
 package roomescape;
 
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
+import roomescape.config.WebConfig;
 
-@SpringBootApplication
 public class RoomescapeApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(RoomescapeApplication.class, args);
+        SpringApplication.run(WebConfig.class, args);
     }
 }
