+++ src/main/java/roomescape/config/ConsoleConfig.java
@@ -1,0 +1,12 @@
+package roomescape.config;
+
+import org.springframework.context.annotation.ComponentScan;
+
+@ComponentScan(basePackages = {
+        "roomescape.controller.console",
+        "roomescape.repository.memory",
+        "roomescape.service",
+        "roomescape.view"})
+public class ConsoleConfig {
+
+}
