+++ src/main/java/roomescape/config/WebConfig.java
@@ -1,0 +1,14 @@
+package roomescape.config;
+
+import org.springframework.boot.SpringBootConfiguration;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+
+@SpringBootConfiguration
+@EnableAutoConfiguration
+@ComponentScan(basePackages = {
+        "roomescape.controller.web",
+        "roomescape.repository.h2",
+        "roomescape.service",})
+public class WebConfig {
+}
