+++ src/main/java/roomescape/repository/h2/ReservationTimeDAO.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.repository.h2;
 
 import java.util.HashMap;
 import java.util.List;
@@ -8,6 +8,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.model.ReservationTime;
+import roomescape.repository.ReservationTimeRepository;
 
 @Repository
 public class ReservationTimeDAO implements ReservationTimeRepository {
