+++ src/test/java/roomescape/repository/ReservationMemoryRepositoryTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import roomescape.model.Reservation;
 import roomescape.model.ReservationTime;
+import roomescape.repository.memory.ReservationMemoryRepository;
 
 
 class ReservationMemoryRepositoryTest {
